+++ NEWS.md
@@ -2,6 +2,7 @@
 - released: ???
 - implemented sparse bit chain for fast computation of dig() on sparse crisp
   data
+- added clustering characteristics to explore() for association rules
 
 # nuggets 2.2.1
 - released: 2026-06-10